spatial division

+++ raytracer/output.py
@@ -80,7 +80,9 @@ if has_PIL:
                 colour[3] = 0
 
             self.__pixels__[x, y] = (
-                colour[1] * 255, colour[2] * 255, colour[3] * 255)
+                int(colour[1] * 255),
+                int(colour[2] * 255),
+                int(colour[3] * 255))
 
         def get_output(self):
             """Returns the PIL image created"""
